% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chrisEx1}
\alias{chrisEx1}
\title{Randomised Block design consisted of 6 blocks and 3 plots.}
\format{
A data frame with 18 rows and 5 variables: \describe{
\item{Blocks}{Block factor containing 6 levels} \item{Plots}{Plot factor
containing 3 levels}
\item{A}{Treatment factor A containing 3 levels}
\item{B}{Treatment factor B containing 3 levels} 
\item{C}{Treatment factor C containing 9 levels} }
}
\usage{
chrisEx1
}
\description{
Randomised Block design consisted of 6 blocks and 3 plots.
}
\examples{
 data(chrisEx1)

summaryAovOnePhase(chrisEx1, "Blocks/Plots", "A*B*C")
}
\keyword{datasets}
