% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chrisEx3}
\alias{chrisEx3}
\title{Randomised Block design consisted of 4 blocks and 2 plots.}
\format{
A data frame with 8 rows and 5 variables: \describe{
\item{Blocks}{Block factor containing 4 levels} \item{Plots}{Plot factor
containing 2 levels} \item{A}{Treatment factor A containing 2 levels}
\item{B}{Treatment factor B containing 2 levels} \item{C}{Treatment factor
C containing 4 levels} }
}
\usage{
chrisEx3
}
\description{
Randomised Block design consisted of 4 blocks and 2 plots.
}
\examples{
 data(chrisEx2)

summaryAovOnePhase(chrisEx2, "Blocks/Plots", "A*B*C") 
}
\keyword{datasets}
