% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identityMat.R
\name{identityMat}
\alias{identityMat}
\title{Identity Matrix}
\usage{
identityMat(n)
}
\arguments{
\item{n}{a numeric describes the dimension of the identity matrix.}
}
\value{
This function returns a matrix with the diagonal elements equal to
one and the off-diagonal elements equal to zero.
}
\description{
Construct an identity matrix.
}
\examples{


identityMat(10)


}
\references{
John J, Williams E (1987). \emph{Cyclic and computer generated
Designs}. Second edition. Chapman & Hall.
}
\seealso{
\code{\link{diag}}
}
\author{
Kevin
}
