% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoDecompuTE-package.R
\docType{package}
\name{infoDecompuTE-package}
\alias{infoDecompuTE-package}
\alias{infoDecompuTE}
\title{Information Decomposition of Two-phase Experiments}
\description{
he main purpose of this package is to generate the structure of the analysis of variance 
(ANOVA) table of the two-phase experiments. The user only need to input the design and the
relationships of the random and fixed factors using the Wilkinson-Rogers' syntax, 
this package can then quickly generate the structure of the ANOVA table with the 
coefficients of the variance components for the expected mean squares.
Thus, the balanced incomplete block design and provides the efficiency
factors of the fixed effects can also be studied and compared much easily.
}
\details{
\tabular{ll}{ Package: \tab infoDecompuTE\cr Type: \tab Package\cr Version:
\tab 0.6.1\cr Date: \tab 2018-05-28\cr License: \tab GPL (>= 3) \cr
LazyLoad: \tab yes\cr }
}
\author{
Kevin Chang and Katya Ruggiero

Maintainer: Kevin Chang <k.chang@auckland.ac.nz>
}
\keyword{package}
