% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J.R
\name{J}
\alias{J}
\title{Identity Matrix Minus Averaging Matrix}
\usage{
J(n)
}
\arguments{
\item{n}{a numeric describes the dimension of the square matrix.}
}
\value{
This function return a square matrix which the identity matrix
minus the averaging matrix.
}
\description{
Construct a square matrix which the identity matrix minus the averging
matrix.
}
\examples{


J(10)


}
\author{
Kevin Chang
}
