% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K.R
\name{K}
\alias{K}
\title{Averaging Matrix}
\usage{
K(n)
}
\arguments{
\item{n}{a numeric describes the dimension of the averaging matrix.}
}
\value{
This function returns a \eqn{n \times n} square matrix with all
elements equal 1/n.
}
\description{
Construct a n-by-n averaging matrix.
}
\examples{


K(10)


}
\references{
John J, Williams E (1987). \emph{Cyclic and computer generated
Designs}. Second edition. Chapman & Hall.
}
\author{
Kevin Chang
}
