\name{m1}
\alias{m1}
\title{
Vector of 1_n
}
\description{
Construct a vector with all elements unity.
}
\usage{
m1(n)
}
\arguments{
  \item{n}{
a numeric describe the length of vector.
}
}

\value{
This function returns a n-by-1 matrix will all elements unity.
}
\references{
John J, Williams E (1987). \emph{Cyclic and computer generated Designs}. Second edition. Chapman
& Hall.
}
\author{
Kevin Chang
}
\examples{
m1(10)
}
