\name{mJ}
\alias{mJ}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Identity Matrix Minus Averging Matrix
}
\description{
Construct a square matrix which the identity matrix minus the averging matrix.
}
\usage{
mJ(n)
}
\arguments{
  \item{n}{
a numeric describes the dimension of the square matrix.
}
}
\value{
This function return a square matrix which the identity matrix minus the averging matrix.
}
\author{
Kevin Chang
}

\examples{
mJ(10)
}
