% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projMat.R
\name{projMat}
\alias{projMat}
\title{Construct a Projection Matrix}
\usage{
projMat(X)
}
\arguments{
\item{X}{a square matrix.}
}
\value{
A square matrix.
}
\description{
Compute the projection matrix from a square matrix.
}
\examples{


m = matrix(1, nrow = 10, ncol = 3)
projMat(m) 


}
\author{
Kevin Chang
}
