% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr.R
\name{tr}
\alias{tr}
\title{Trace of the Matrix}
\usage{
tr(X)
}
\arguments{
\item{X}{a square matrix.}
}
\value{
A numeric value.
}
\description{
Compute the trace of the square matrix.
}
\examples{


m = matrix(1, nrow = 10, ncol = 10)
tr(m)   


}
\references{
John J, Williams E (1987). \emph{Cyclic and computer generated
Designs}. Second edition. Chapman & Hall.
}
\seealso{
\code{\link{diag}}
}
\author{
Kevin
}
