% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unity.R
\name{unity}
\alias{unity}
\title{Construct a unity vector}
\usage{
unity(n)
}
\arguments{
\item{n}{a numeric describe the length of vector.}
}
\value{
This function returns a \eqn{n \\times 1} matrix will all elements
unity.
}
\description{
Construct a vector with all elements unity.
}
\examples{


unity(10)


}
\references{
John J, Williams E (1987). \emph{Cyclic and computer generated
Designs}. Second edition. Chapman & Hall.
}
\author{
Kevin Chang
}
