\name{adjustEffectNames}
\alias{adjustEffectNames}
\title{
Adjust the Effects' Names
}
\description{
Adjust for appropriate syntax describing the effects matching the structual formula.
}
\usage{
adjustEffectNames(str, effectsMatrix, effectNames)
}
\arguments{
\item{str}{
a single string of characters containing the structural formula for the block factors using the Wilkinson-Rogers' syntax.
}
\item{effectNames}{
a vector of charactor containing the labels of the treatment or block terms in the model generated by the \code{\link{terms}}.
}
\item{effectsMatrix}{
a matrix of variables by terms showing which variables appear in which terms generated by the \code{\link{terms}}.
}
}

\value{
A vector of charactor containing the labels of the terms in the model with appropriate syntax describing the effects. 
}
\author{
Kevin Chang
}
\examples{
str = "A*(B/E/C)*D"
effectsMatrix= attr(terms(as.formula(paste("~", str)), keep.order = TRUE) , "factor")
effectNames =  attr(terms(as.formula(paste("~", str)), keep.order = TRUE) , "term.labels")

adjustEffectNames(str, effectsMatrix, effectNames) 
}
