\name{getEffFactor}
\alias{getEffFactor}
\title{
Compute the Efficiency Factors of Treatment effects
}
\description{
Compute the Efficiency Factors for every treatment effect within a single stratum.
}
\usage{
getEffFactor(z, T, N, Rep)
}
\arguments{
   \item{z}{
a matrix containing the orthogonal projector of a stratum generated by \code{\link{makeOrthProjectors}}.
}
  \item{T}{
a list of contrast matrices generated by \code{\link{makeContrMat}}.
}
  \item{N}{
a matrix containing the design matrix generated by \code{\link{makeOverDesMat}}.
}     
   \item{Rep}{
a numeric vector containing the treatment replication number generated by \code{\link{getTrtRep}} 
}
}
\value{
A list of numeric vectors containing the efficiency factors of every treatment effect.
}
\details{
The main purpose of this function is to compute the canoncial or average efficiency factors for each treatment effects in each stratum of ANOVA table. 

The canoncial efficiency factors are generated when the user input the treatment contrasts, otherwise the average efficiency factors, which is the harmomic mean of the canoncial efficiency factors, are gnerated.  
}

\author{
Kevin Chang
}
\examples{
design1 <- local({ 
  Ani = as.factor(LETTERS[c(1,2,3,4,
                            5,6,7,8)])
  Trt = as.factor(letters[c(1,1,1,1,
                            2,2,2,2)])
  data.frame(Ani, Trt)
})

blk.str = "Ani"
    
rT = terms(as.formula(paste("~", blk.str, sep = "")), keep.order = TRUE) 
blkTerm = attr(rT,"term.labels")
     
Z = makeBlkDesMat(design1, blkTerm)

trt.str = "Trt"              
fT <- terms(as.formula(paste("~", trt.str, sep = "")), keep.order = TRUE)  #fixed terms

trtTerm <- attr(fT, "term.labels")
effectsMatrix <- attr(fT, "factor")        

T <- makeContrMat(design1, trtTerm, effectsMatrix, contr.vec = NA , 
        contr.matrix = TRUE)

N =  makeOverDesMat(design1, trtTerm)

Rep = getTrtRep(design1, trtTerm)   
 
effFactors = lapply(makeOrthProjectors(Z), function(z) getEffFactor(z, T, N, Rep))

}
