\name{infoDecompuTE-package}
\alias{infoDecompuTE-package}
\alias{infoDecompuTE}
\docType{package}
\title{
Computation of the Coefficients of the Variance Components for the Expected Mean Squares 
}
\description{         
infoDecompuTE is a package designed specifically for studying the analysis of variance (ANOVA) 
table of the two-phase experiments. By inputting the design and the relationships of the random 
and fixed factors using the Wilkinson-Rogers syntax, infoDecompuTE can generate the structure of 
the ANOVA table with the coefficients of the variance components for the expected mean squares. 
This package can also study the balance incomplete block design and provides the efficiency }
\details{
\tabular{ll}{
Package: \tab infoDecompuTE\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-02-28\cr
License: \tab GPL-3 \cr
LazyLoad: \tab yes\cr
}
}
\author{
Kevin Chang and Katya Ruggiero

Maintainer: Kevin Chang <kcha193@aucklanduni.ac.nz> 
}

\keyword{ package }
