\name{infoDecompuTE-package}
\alias{infoDecompuTE-package}
\alias{infoDecompuTE}
\docType{package}
\title{
Information Decomposition of Two-phase Experiments
}
\description{         
InfoDecompuTE is capable of generating the structure of the analysis of variance (ANOVA) table of the two-phase experiments. By inputting the design and the relationships of the random and fixed factors using the Wilkinson-Rogers' syntax, infoDecompuTE can generate the structure of the ANOVA table with the coefficients of the variance components for the expected mean squares. This package can also study the balanced incomplete block design and provides the efficiency factors of the fixed effects. 
}
\details{
\tabular{ll}{
Package: \tab infoDecompuTE\cr
Type: \tab Package\cr
Version: \tab 0.5.1\cr
Date: \tab 2013-01-04\cr
License: \tab GPL (>= 3) \cr
LazyLoad: \tab yes\cr
}
}
\author{
Kevin Chang and Katya Ruggiero

Maintainer: Kevin Chang <kcha193@aucklanduni.ac.nz> 
}

\keyword{package}
