\name{J}
\alias{J}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Identity Matrix Minus Averaging Matrix
}
\description{
Construct a square matrix which the identity matrix minus the averaging matrix.
}
\usage{
J(n)
}
\arguments{
  \item{n}{
a numeric describes the dimension of the square matrix.
}
}
\value{
This function return a square matrix which the identity matrix minus the averaging matrix.
}
\author{
Kevin Chang
}

\examples{
J(10)
}
