\name{unity}
\alias{unity}
\title{
Construct a unity vector 
}
\description{
Construct a vector with all elements unity.
}
\usage{
unity(n)
}
\arguments{
  \item{n}{
a numeric describe the length of vector.
}
}

\value{
This function returns a \eqn{n \\times 1} matrix will all elements unity.
}
\references{
John J, Williams E (1987). \emph{Cyclic and computer generated Designs}. Second edition. Chapman
& Hall.
}
\author{
Kevin Chang
}
\examples{
unity(10)
}
